fix: relocate parseName to validators file
